Fix git commit failures in worktrees by writing temp files to worktree root
The issue was that Crystal was trying to write temporary commit message files to .git/FILENAME, but in git worktrees, .git is a file (not a directory) that contains a pointer to the actual git directory.

Changes:
- commitManager.ts: Write COMMIT_MSG_TEMP to worktree root instead of .git/
- worktreeManager.ts: Write SQUASH_MSG to worktree root instead of .git/

This fixes the ENOENT error when Crystal tries to create commits in worktrees.